Move db scripts under the database workspace
