+++ .eslintrc
@@ -16,7 +16,8 @@
           "**/packages/psss/src/api/FakeApi.js",
           "**/packages/psss/src/utils/test-utils.js",
           "**/packages/ui-components/helpers/testingRenderer.js",
-          "**/packages/utils/src/script.js",
+          "**/packages/utils/src/runScript.js",
+          "**/packages/utils/src/Script.js",
           "**/packages/web-frontend/config/**",
           "**/packages/web-frontend/scripts/**"
         ],
@@ -177,7 +178,7 @@
         "react/state-in-constructor": "warn",
         "react/static-property-placement": "warn",
         "prefer-destructuring": "warn",
-        "prettier/prettier": "warn",
+        "prettier/prettier": "warn"
       }
     }
   ]
