+++ package.json
@@ -22,13 +22,13 @@
     "build-internal-dependencies": "./scripts/bash/buildInternalDependencies.sh",
     "build-internal-dependencies:dev": "./scripts/bash/buildInternalDependencies.sh --withTypes",
     "build-web-frontend": "yarn workspace @tupaia/web-frontend build",
-    "dump-database": "./scripts/bash/dumpDb.sh",
+    "dump-database": "yarn workspace @tupaia/database dump-database",
     "migrate": "yarn workspace @tupaia/database migrate",
     "migrate-create": "yarn workspace @tupaia/database migrate-create",
     "migrate-down": "yarn workspace @tupaia/database migrate-down",
     "package-json-lint": "npmPkgJsonLint --configFile .npmpackagejsonlintrc.json .",
     "postinstall": "test -n \"$SKIP_BUILD_INTERNAL_DEPENDENCIES\" || yarn build-internal-dependencies",
-    "refresh-database": "node scripts/node/refreshDatabase.js",
+    "refresh-database": "yarn workspace @tupaia/database refresh-database",
     "test-all": "node ./scripts/node/testAll --silent",
     "update-codeship-env-vars": "./packages/devops/scripts/ci/updateCiEnvVars.sh",
     "download-parameter-store-env-vars": "./scripts/bash/syncDotEnvWithParameterStore.sh --download",
