+++ packages/database/scripts/refreshDatabase.js
@@ -5,7 +5,7 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
-const Script = require('./Script');
+const { Script } = require('@tupaia/utils');
 
 class RefreshDatabaseScript extends Script {
   config = {
