+++ packages/utils/src/index.js
@@ -27,7 +27,8 @@ export * from './object';
 export * from './request';
 export { replaceValues } from './replaceValues';
 export { respond } from './respond';
-export * from './script';
+export * from './runScript';
+export * from './Script';
 export * from './string';
 export * from './validation';
 export { WorkBookParser } from './WorkBookParser';
