+++ scripts/node/validateTests.js
@@ -9,7 +9,6 @@ const fs = require('fs');
 const path = require('path');
 
 const { getLoggerInstance } = require('@tupaia/utils');
-const { executeCommand } = require('./utilities');
 
 /**
  * @typedef {Array<{ fileName, text }>} ExclusiveTest
@@ -19,6 +18,11 @@ const PACKAGE_ROOT = path.resolve(`${__dirname}/../../packages`);
 const TEST_PATHS_RELATIVE_TO_PACKAGE = ['src/tests', 'src/__tests__', 'tests', '__tests__'];
 const MOCHA_TEST_BLOCKS = ['it', 'describe', 'context', 'specify'];
 const STD_ERROR_CODE = 1;
+
+const executeCommand = async command => {
+  const { stdout } = await promisify(exec)(command);
+  return stdout;
+};
 
 const findTestDirs = () => {
   const packageNames = fs.readdirSync(`${PACKAGE_ROOT}`);
